Refactor files and folder names match to our convention
