+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import { Navigation } from './components/shared';
-import { Footer } from './components/mainPage/Footer/Footer';
+import { Footer } from './components/mainPage/footer/Footer';
 import Views from './Views';
 
 function App() {
