+++ src/components/mainPage/banner/Banner.jsx
