+++ src/components/mainPage/banner/style.css
