+++ src/components/mainPage/footer/style.css
