+++ src/components/shared/api/Api.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
-const API = axios.create({
+const Api = axios.create({
     baseURL: process.env.REACT_APP_API_URL,
 });
 
-export default API;
+export default Api;
